requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
tabulate>=0.9.0
pytest>=7.0.0
